backend/requirements.txt
flask
flask-cors
flask-mysqldb
python-dotenv
